# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sahara Event Core API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EventStatus {
  id: string;
  name: string;
  stage: string;
  homeTeam: string;
  awayTeam: string;
  homeScore: number;
  awayScore: number;
  minute: number;
  isLive: boolean;
  possession: number;
  atmosphereDb: number;
  weatherCelsius: number;
  userSection: string;
  userRow: string;
  userSeat: string;
}

export interface DashboardStats {
  totalAttendance: number;
  attendanceChange: number;
  avgGateEntryMinutes: number;
  gateEntryStatus: string;
  activeAlerts: number;
  alertImpact: string;
  systemHealth: number;
}

export type CrowdZoneDensity =
  (typeof CrowdZoneDensity)[keyof typeof CrowdZoneDensity];

export const CrowdZoneDensity = {
  low: "low",
  moderate: "moderate",
  high: "high",
  peak: "peak",
} as const;

export interface CrowdZone {
  id: string;
  name: string;
  level: string;
  density: CrowdZoneDensity;
  percentage: number;
}

export type GateStatusCongestion =
  (typeof GateStatusCongestion)[keyof typeof GateStatusCongestion];

export const GateStatusCongestion = {
  smooth: "smooth",
  moderate: "moderate",
  critical: "critical",
} as const;

export interface GateStatus {
  id: string;
  name: string;
  location: string;
  type: string;
  waitMinutes: number;
  congestion: GateStatusCongestion;
  passengersPerMin: number;
  fillPercent: number;
}

export type AmenityStatusStatus =
  (typeof AmenityStatusStatus)[keyof typeof AmenityStatusStatus];

export const AmenityStatusStatus = {
  vacant: "vacant",
  moderate: "moderate",
  at_capacity: "at_capacity",
} as const;

export interface AmenityStatus {
  id: string;
  name: string;
  location: string;
  status: AmenityStatusStatus;
}

export interface CongestionForecast {
  gate: string;
  forecastText: string;
  expectedIncreasePercent: number;
  expectedAtTime: string;
  reason: string;
  validatedBy: string;
}

export type VendorWaitTrend =
  (typeof VendorWaitTrend)[keyof typeof VendorWaitTrend];

export const VendorWaitTrend = {
  decreasing: "decreasing",
  stable: "stable",
  increasing: "increasing",
} as const;

export interface Vendor {
  id: string;
  name: string;
  category: string;
  zone: string;
  rating: number;
  waitMinutes: number;
  waitTrend: VendorWaitTrend;
  isInstantPickup: boolean;
  imageUrl?: string;
}

export interface MenuItem {
  id: string;
  name: string;
  price: number;
  description?: string;
}

export type VendorDetail = Vendor & {
  description?: string;
  menuItems: MenuItem[];
};

export interface PickupWindow {
  timeLabel: string;
  busynessLevel: number;
  isOptimal: boolean;
}

export type PlaceOrderRequestItemsItem = {
  menuItemId: string;
  quantity: number;
};

export interface PlaceOrderRequest {
  vendorId: string;
  items: PlaceOrderRequestItemsItem[];
  pickupTime?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  collected: "collected",
} as const;

export interface Order {
  id: string;
  vendorId: string;
  vendorName: string;
  items: string[];
  status: OrderStatus;
  pickupLocation: string;
  estimatedMinutes: number;
  createdAt: string;
}

export type NotificationPriority =
  (typeof NotificationPriority)[keyof typeof NotificationPriority];

export const NotificationPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Notification {
  id: string;
  priority: NotificationPriority;
  title: string;
  message: string;
  timestamp: string;
  isDismissed: boolean;
  actionLabel?: string;
  actionRoute?: string;
}

export interface UserProfile {
  id: string;
  name: string;
  guestId: string;
  tier: string;
  section: string;
  row: string;
  seat: string;
  venue: string;
  wheelchairRoutes: boolean;
  assistiveListening: boolean;
  reducedSensoryZones: boolean;
  topicInterests: string[];
  cateringPreferences: string[];
}

export interface UpdateProfileRequest {
  wheelchairRoutes?: boolean;
  assistiveListening?: boolean;
  reducedSensoryZones?: boolean;
  topicInterests?: string[];
  cateringPreferences?: string[];
}

export type GetVendorsParams = {
  category?: string;
};