# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sahara Event Core API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders/vendors:
    get:
      operationId: getVendors
      tags: [orders]
      summary: Get all food vendors
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vendor list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
  /orders/vendors/{vendorId}:
    get:
      operationId: getVendorById
      tags: [orders]
      summary: Get vendor details and menu
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vendor with menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetail"
...
  /orders:
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place a food order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get current user orders
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Vendor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        zone:
          type: string
        rating:
          type: number
        waitMinutes:
          type: integer
        waitTrend:
          type: string
          enum: [decreasing, stable, increasing]
        isInstantPickup:
          type: boolean
        imageUrl:
          type: string
      required:
        - id
        - name
        - category
        - zone
        - rating
        - waitMinutes
        - waitTrend
        - isInstantPickup
    VendorDetail:
      allOf:
        - $ref: "#/components/schemas/Vendor"
        - type: object
          properties:
            description:
              type: string
            menuItems:
              type: array
              items:
                $ref: "#/components/schemas/MenuItem"
          required:
            - menuItems
    MenuItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        description:
          type: string
      required:
        - id
        - name
        - price
...
    PlaceOrderRequest:
      type: object
      properties:
        vendorId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: string
              quantity:
                type: integer
            required:
              - menuItemId
              - quantity
        pickupTime:
          type: string
      required:
        - vendorId
        - items
    Order:
      type: object
      properties:
        id:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        items:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, preparing, ready, collected]
        pickupLocation:
          type: string
        estimatedMinutes:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - vendorId
        - vendorName
        - items
        - status
        - pickupLocation
        - estimatedMinutes
        - createdAt
...