# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sahara Event Core API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: string;
  name: string;
  price: number;
  description?: string;
}