# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sahara Event Core API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  vendorId: string;
  vendorName: string;
  items: string[];
  status: OrderStatus;
  pickupLocation: string;
  estimatedMinutes: number;
  createdAt: string;
}