# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sahara Event Core API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VendorWaitTrend } from "./vendorWaitTrend";

export interface Vendor {
  id: string;
  name: string;
  category: string;
  zone: string;
  rating: number;
  waitMinutes: number;
  waitTrend: VendorWaitTrend;
  isInstantPickup: boolean;
  imageUrl?: string;
}